size_t problem on Mac

+++ src/idl_parser.cpp
@@ -1461,7 +1461,7 @@ void SimpleQsort(T *begin, T *end, size_t width, F comparator, S swapper) {
 CheckedError Parser::ParseAlignAttribute(const std::string &align_constant,
                                          size_t min_align, size_t *align) {
   // Use small temporary to avoid possible problems with size_t on Mac.
-  uint8_t align_value;
+  size_t align_value;
   if (StringToNumber(align_constant.c_str(), &align_value) &&
       VerifyAlignmentRequirements(static_cast<size_t>(align_value),
                                   min_align)) {
